requirements.txt
pandas
faker
numpy
duckdb
plotly
streamlit
pyyaml
fastapi
uvicorn
python-multipart
requests
pytest
httpx
watchdog
dbt-duckdb